Move classification helpers to reduce validate.js file length
