+++ src/application/summary-logs/load-counts.js
@@ -1,10 +1,17 @@
 import { VERSION_STATUS } from '#domain/waste-records/model.js'
 import { ROW_OUTCOME } from '#domain/summary-logs/table-schemas/validation-pipeline.js'
-import { PROCESSING_TYPE_TABLES } from '#domain/summary-logs/table-schemas/index.js'
+import {
+  PROCESSING_TYPE_TABLES,
+  findSchemaForProcessingType
+} from '#domain/summary-logs/table-schemas/index.js'
+import { isNil } from '#common/helpers/is-nil.js'
 import { SUMMARY_LOG_STATUS } from '#domain/summary-logs/status.js'
+import { computeLoadsByPeriodStatus } from './period-status.js'
 
 /** @import {ValidatedWasteRecord} from '#application/waste-records/transform-from-summary-log.js' */
 /** @import {ValidationIssue} from '#common/validation/validation-issues.js' */
+/** @import {TypedLogger} from '#common/helpers/logging/logger.js' */
+/** @import {ReportsRepository} from '#reports/repository/port.js' */
 
 /**
  * @typedef {Object} LoadCategory
@@ -298,4 +305,89 @@ export const classifyLoads = ({
   })
 
   return { loads, loadsByWasteRecordType }
+}
+
+/**
+ * Filters waste records to only those from tables that participate in waste balance.
+ *
+ * @param {ValidatedWasteRecord[] | null} wasteRecords
+ * @param {string} [processingType]
+ * @returns {ValidatedWasteRecord[]}
+ */
+const filterWasteBalanceRecords = (wasteRecords, processingType) => {
+  if (!processingType) {
+    return []
+  }
+  return (
+    wasteRecords?.filter((wr) => {
+      const schema = findSchemaForProcessingType(processingType, wr.record.type)
+      return !isNil(schema?.classifyForWasteBalance)
+    }) ?? []
+  )
+}
+
+/**
+ * Classifies loads and computes period status for a validated summary log.
+ *
+ * @param {Object} params
+ * @param {ValidatedWasteRecord[] | null} params.wasteRecords
+ * @param {string} params.summaryLogId
+ * @param {string} params.status
+ * @param {import('#domain/summary-logs/meta-fields.js').ProcessingType} [params.processingType]
+ * @param {import('#domain/organisations/registration.js').Registration} [params.registration]
+ * @param {Map<string, import('#domain/waste-records/model.js').WasteRecord>} [params.existingRecordsMap]
+ * @param {ReportsRepository} params.reportsRepository
+ * @param {string} params.organisationId
+ * @param {string} params.registrationId
+ * @param {string} params.loggingContext
+ * @param {TypedLogger} params.logger
+ */
+export const classifyAndComputePeriodStatus = async ({
+  wasteRecords,
+  summaryLogId,
+  status,
+  processingType,
+  registration,
+  existingRecordsMap,
+  reportsRepository,
+  organisationId,
+  registrationId,
+  loggingContext,
+  logger
+}) => {
+  const wasteBalanceRecords = filterWasteBalanceRecords(
+    wasteRecords,
+    processingType
+  )
+
+  const { loads, loadsByWasteRecordType } = classifyLoads({
+    processingType,
+    status,
+    summaryLogId,
+    wasteBalanceRecords,
+    wasteRecords
+  })
+
+  const loadsByPeriodStatus = await computeLoadsByPeriodStatus({
+    wasteRecords,
+    wasteBalanceRecords,
+    summaryLogId,
+    registration,
+    processingType,
+    existingRecordsMap,
+    reportsRepository,
+    status,
+    organisationId,
+    registrationId,
+    loggingContext,
+    logger,
+    processingTypeTables: PROCESSING_TYPE_TABLES
+  })
+
+  return {
+    loads,
+    loadsByWasteRecordType,
+    loadsByPeriodStatus,
+    wasteBalanceRecords
+  }
 }
